split the massive FinishGen.cpp
