+++ src/Generating/Finishers/CMakeLists.txt
@@ -1,0 +1,26 @@
+target_sources(
+        ${CMAKE_PROJECT_NAME} PRIVATE
+    FinishGen.h
+    FinishGenClumpTopBlock.h
+    FinishGenFluidSprings.h
+    FinishGenGlowStone.h
+    FinishGenNetherClumpFoliage.h
+    FinishGenOres.h
+    FinishGenPassiveMobs.h
+    FinishGenPreSimulator.h
+    FinishGenSingleTopBlock.h
+    FinishGenSprinkleFoliage.h
+    FinishGenTallGrass.h
+
+    FinishGen.cpp
+    FinishGenClumpTopBlock.cpp
+    FinishGenFluidSprings.cpp
+    FinishGenGlowStone.cpp
+    FinishGenNetherClumpFoliage.cpp
+    FinishGenOres.cpp
+    FinishGenPassiveMobs.cpp
+    FinishGenPreSimulator.cpp
+    FinishGenSingleTopBlock.cpp
+    FinishGenSprinkleFoliage.cpp
+    FinishGenTallGrass.cpp
+)
