+++ src/Generating/Finishers/FinishGenTallGrass.h
@@ -1,0 +1,22 @@
+
+// FinishGenTallGrass.h
+
+#pragma once
+
+class cFinishGenTallGrass :
+	public cFinishGen
+{
+public:
+	cFinishGenTallGrass(int a_Seed) : m_Noise(a_Seed), m_Seed(a_Seed) {}
+
+protected:
+	cNoise m_Noise;
+	int m_Seed;
+
+	// cFinishGen override:
+	virtual void GenFinish(cChunkDesc & a_ChunkDesc) override;
+
+	static bool CanFernGrow(EMCSBiome a_Biome);
+	static bool CanLargeFernGrow(EMCSBiome a_Biome);
+	static int GetBiomeDensity(EMCSBiome a_Biome);
+};
